apps/api/requirements.txt
Package             Version
------------------- -------
asgiref             3.8.1
Django              5.0.4
djangorestframework 3.15.1
pip                 22.0.2
psycopg             3.1.18
psycopg-binary      3.1.18
python-dotenv       1.0.1
setuptools          59.6.0
sqlparse            0.5.0
typing_extensions   4.11.0